Refactor DisplayIdentity class, add iterface to it.
